Setup Level1 and Level2 also linked to main menu
Starting from main menu now will take you to Level 1 map which is the forest map. Win condition will be collecting 3 peices of a map(object) while constantly being attacked by monsters. After 3 peices of a map are found player cutscene is played and screen is moved to Scence 2 (Level 2) which is the desert sand map.